moca2024: best firmware hack commit
Actually this commit disable the wifi and bt initialization (more cleanup needed); add the business login to randomize the 'magic' answer; add some images as 'magic' answers (later will be converted to text and may be implemented a new feature :)); change the startup logo to something appropriate for the event :). Thanks for the Flipper Zero, thanks for all. Sharing is caring.